Improve handling ballerina source files inside non-ballerina projects
