+++ tool-plugins/intellij/src/main/java/io/ballerina/plugins/idea/preloading/BallerinaPreloadingActivity.java
@@ -15,36 +15,20 @@
  */
 package io.ballerina.plugins.idea.preloading;
 
-import com.google.common.base.Strings;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.PreloadingActivity;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.editor.EditorFactory;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManager;
 import com.intellij.openapi.project.ProjectManagerListener;
-import com.intellij.openapi.ui.MessageType;
 import com.intellij.util.messages.MessageBusConnection;
-import io.ballerina.plugins.idea.extensions.BallerinaLSPExtensionManager;
-import io.ballerina.plugins.idea.notifiers.BallerinaAutoDetectNotifier;
-import io.ballerina.plugins.idea.sdk.BallerinaSdk;
 import io.ballerina.plugins.idea.sdk.BallerinaSdkUtils;
-import io.ballerina.plugins.idea.settings.autodetect.BallerinaAutoDetectionSettings;
-import io.ballerina.plugins.idea.settings.langserverlogs.LangServerLogsSettings;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.wso2.lsp4intellij.IntellijLanguageClient;
-import org.wso2.lsp4intellij.client.languageserver.serverdefinition.ProcessBuilderServerDefinition;
 
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
-
-import static io.ballerina.plugins.idea.BallerinaConstants.BAL_FILE_EXT;
-import static io.ballerina.plugins.idea.BallerinaConstants.LAUNCHER_SCRIPT_PATH;
-import static io.ballerina.plugins.idea.BallerinaConstants.SYS_PROP_EXPERIMENTAL;
-import static io.ballerina.plugins.idea.BallerinaConstants.SYS_PROP_LS_DEBUG;
-import static io.ballerina.plugins.idea.BallerinaConstants.SYS_PROP_LS_TRACE;
+import static io.ballerina.plugins.idea.preloading.LSPUtils.registerServerDefinition;
 import static io.ballerina.plugins.idea.preloading.OSUtils.getOperatingSystem;
 
 /**
@@ -52,8 +36,7 @@ import static io.ballerina.plugins.idea.preloading.OSUtils.getOperatingSystem;
  */
 public class BallerinaPreloadingActivity extends PreloadingActivity {
 
-    private static final Logger LOG = Logger.getInstance(BallerinaPreloadingActivity.class);
-    private static BallerinaAutoDetectNotifier autoDetectNotifier = new BallerinaAutoDetectNotifier();
+    static final Logger LOG = Logger.getInstance(BallerinaPreloadingActivity.class);
 
     /**
      * Preloading of the ballerina plugin.
@@ -70,7 +53,20 @@ public class BallerinaPreloadingActivity extends PreloadingActivity {
         connect.subscribe(ProjectManager.TOPIC, new ProjectManagerListener() {
             @Override
             public void projectOpened(@Nullable final Project project) {
-                registerServerDefinition(project);
+                if (project == null) {
+                    return;
+                }
+                if (isBallerinaProject(project)) {
+                    // If the opened project root is a ballerina project root, proceeds with SDK validation and auto
+                    // detection, if needed.
+                    registerServerDefinition(project);
+                } else {
+                    // Even if the project is not a ballerina project(does not have ballerina project structure),
+                    // it might contains ballerina modules or single ballerina source files. Therefore an editor
+                    // listener is used to trigger the language server connection and proceed accordingly, when it
+                    // detects a ballerina file being opened.
+                    watchForBalFiles(project);
+                }
             }
         });
 
@@ -81,99 +77,6 @@ public class BallerinaPreloadingActivity extends PreloadingActivity {
             }
             return true;
         });
-    }
-
-    /**
-     * Registered language server definition using currently opened ballerina projects.
-     */
-    private static void registerServerDefinition() {
-        Project[] openProjects = ProjectManager.getInstance().getOpenProjects();
-        for (Project project : openProjects) {
-            registerServerDefinition(project);
-        }
-    }
-
-    private static boolean registerServerDefinition(Project project) {
-
-        boolean autoDetected = false;
-
-        //If the project does not have a ballerina SDK attached, ballerinaSdkPath will be null.
-        BallerinaSdk balSdk = BallerinaSdkUtils.getBallerinaSdkFor(project);
-        String balSdkPath = balSdk.getSdkPath();
-
-        // Checks for the user-configured auto detection settings.
-        if (balSdkPath == null && BallerinaAutoDetectionSettings.getInstance(project).isAutoDetectionEnabled()) {
-
-            //If a ballerina SDK is not configured for the project, Plugin tries to auto detect the ballerina SDK.
-            showInIdeaEventLog(project, String.format("No ballerina SDK is found for project: %s\n " +
-                    "Trying to Auto detect Ballerina Home...", project.getBasePath()));
-
-            balSdkPath = BallerinaSdkUtils.autoDetectSdk(project);
-            autoDetected = true;
-        }
-
-        if (!Strings.isNullOrEmpty(balSdkPath)) {
-            boolean success = doRegister(project, balSdkPath);
-            if (success && autoDetected) {
-                LOG.info(String.format("Auto-detected Ballerina Home: %s for the project: %s", balSdkPath,
-                        project.getBasePath()));
-                showInIdeaEventLog(project, "Auto-Detected Ballerina Home: " + balSdkPath);
-            }
-            return success;
-        } else {
-            if (BallerinaAutoDetectionSettings.getInstance(project).isAutoDetectionEnabled()) {
-                showInIdeaEventLog(project, "Auto-Detection Failed for: " + project.getBasePath());
-            }
-        }
-        return false;
-    }
-
-    private static boolean doRegister(@NotNull Project project, @NotNull String sdkPath) {
-        String os = OSUtils.getOperatingSystem();
-        if (os == null) {
-            return false;
-        }
-
-        // Creates the args list to register the language server definition using the ballerina lang-server launcher
-        // script.
-        ProcessBuilder processBuilder;
-        List<String> args = new ArrayList<>();
-        if (os.equals(OSUtils.UNIX) || os.equals(OSUtils.MAC)) {
-            args.add(Paths.get(sdkPath, LAUNCHER_SCRIPT_PATH, "language-server-launcher.sh").toString());
-        } else if (os.equals(OSUtils.WINDOWS)) {
-            args.add(Paths.get(sdkPath, LAUNCHER_SCRIPT_PATH, "language-server-launcher.bat").toString());
-        }
-
-        processBuilder = new ProcessBuilder(args);
-        // Checks user-configurable setting for allowing ballerina experimental features and sets the flag accordingly.
-        if (BallerinaAutoDetectionSettings.getInstance(project).isAutoDetectionEnabled()) {
-            processBuilder.environment().put(SYS_PROP_EXPERIMENTAL, "true");
-        }
-
-        // Checks user-configurable setting for allowing language server debug logs and sets the flag accordingly.
-        if (LangServerLogsSettings.getInstance(project).isLangServerDebugLogsEnabled()) {
-            processBuilder.environment().put(SYS_PROP_LS_DEBUG, "true");
-        }
-
-        // Checks user-configurable setting for allowing language server trace logs and sets the flag accordingly.
-        if (LangServerLogsSettings.getInstance(project).isLangServerTraceLogsEnabled()) {
-            processBuilder.environment().put(SYS_PROP_LS_TRACE, "true");
-        }
-
-        // Adds ballerina-specific custom LSP extensions by creating a ballerina lsp extension manager.
-        IntellijLanguageClient.addExtensionManager(BAL_FILE_EXT, new BallerinaLSPExtensionManager());
-
-        // Registers language server definition in the lsp4intellij lang-client library.
-        IntellijLanguageClient
-                .addServerDefinition(new ProcessBuilderServerDefinition(BAL_FILE_EXT, processBuilder), project);
-
-        LOG.info("language server definition is registered using sdk path: " + sdkPath);
-        return true;
-    }
-
-    private static void showInIdeaEventLog(Project project, String message) {
-        ApplicationManager.getApplication().invokeLater(() -> autoDetectNotifier.showMessage(project, message,
-                MessageType.INFO));
     }
 
     /**
@@ -196,4 +99,18 @@ public class BallerinaPreloadingActivity extends PreloadingActivity {
             LOG.error("Error occurred when trying to terminate ballerina processes", e);
         }
     }
+
+    private void watchForBalFiles(Project project) {
+        EditorFactory.getInstance().addEditorFactoryListener(new BallerinaEditorFactoryListener(project), project);
+    }
+
+    /**
+     * Validates if a given project is a ballerina project, using ballerina project structure.
+     *
+     * @param project The IDEA project instance to be validated.
+     */
+    private static boolean isBallerinaProject(Project project) {
+        return !BallerinaSdkUtils.searchForBallerinaProjectRoot(project.getBasePath(), project.getBasePath()).isEmpty();
+    }
+
 }
